+++ ParaView/ParaView/vtkPVImplicitPlaneWidget.h
@@ -70,16 +70,16 @@ public:
   // Description:
   // Center of the plane.
   void SetCenter();
-  virtual void SetCenter(float,float,float);
-  virtual void SetCenterInternal(float,float,float);
-  virtual void SetCenter(float f[3]) { this->SetCenter(f[0], f[1], f[2]); }
+  virtual void SetCenter(double,double,double);
+  virtual void SetCenterInternal(double,double,double);
+  virtual void SetCenter(double f[3]) { this->SetCenter(f[0], f[1], f[2]); }
 
   // Description:
   // The normal to the plane.
   void SetNormal();
-  virtual void SetNormal(float,float,float);
-  virtual void SetNormalInternal(float,float,float);
-  virtual void SetNormal(float f[3]) { this->SetNormal(f[0], f[1], f[2]); }
+  virtual void SetNormal(double,double,double);
+  virtual void SetNormalInternal(double,double,double);
+  virtual void SetNormal(double f[3]) { this->SetNormal(f[0], f[1], f[2]); }
 
   // Description:
   // The input from the input menu is used to place the widget.
@@ -143,10 +143,10 @@ protected:
 
   vtkClientServerID PlaneID;
 
-  float LastAcceptedCenter[3];
-  float LastAcceptedNormal[3];
-  vtkSetVector3Macro(LastAcceptedCenter, float);
-  vtkSetVector3Macro(LastAcceptedNormal, float);
+  double LastAcceptedCenter[3];
+  double LastAcceptedNormal[3];
+  vtkSetVector3Macro(LastAcceptedCenter, double);
+  vtkSetVector3Macro(LastAcceptedNormal, double);
   
   int ReadXMLAttributes(vtkPVXMLElement* element,
                         vtkPVXMLPackageParser* parser);
